Add displaying of multiple graphs at the same time
I mean, I guess it's in now. But it just isn't useful yet. Need
to take a step back, and reevalueate how to make this actually
useful.
I think we might need interactivity or something like that,
the graphs we're currently creating don't have enough space to
be actually useful. It's just noise.